Move bin directory
