+++ bin/generate-reports-index.js
@@ -2,8 +2,8 @@
 
 'use strict';
 
-const { NewmanConfig } = require('../config');
-const { ReportGenerator } = require('../reports');
+const { NewmanConfig } = require('../src/newman/config');
+const { ReportGenerator } = require('../src/newman/reports');
 
 const newmanConfig = NewmanConfig.load();
 const reportGenerator = new ReportGenerator(newmanConfig);
