+++ package.json
@@ -5,8 +5,13 @@
   "author": "Diverta inc.",
   "license": "UNLICENSED",
   "main": "src/index.js",
-  "scripts": {
+  "bin": {
+    "kuroco-newman-generate-reports-index": "bin/generate-reports-index.js",
+    "kuroco-newman-openapi-to-collection": "bin/openapi-to-collection.js",
+    "kuroco-newman-run-all-collections": "bin/run-all-collections.js",
+    "kuroco-newman-run-collection": "bin/run-collection.js"
   },
+  "scripts": {},
   "repository": {
     "type": "git",
     "url": "git+https://github.com/sakaguchi-diverta/kuroco-newman.git"
